requirements.txt
# Unified dependency set for Project A (model) and Project B (FastAPI app)
# For GPU acceleration install torch/torchvision with the matching CUDA wheel, e.g.:
#   pip install torch==2.3.1+cu121 torchvision==0.18.1+cu121 \
#       --extra-index-url https://download.pytorch.org/whl/cu121

# Project B: FastAPI backend and orchestration
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.20.0
httpx==0.27.0
python-multipart==0.0.9
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.5.1
python-jose[cryptography]==3.3.0
Authlib==1.3.1
stripe==10.12.0
kubernetes==29.0.0

# Shared ML stack
numpy==1.26.4
pillow==10.4.0
transformers==4.44.2
datasets==2.21.0
accelerate==0.33.0
# Default to CPU PyTorch; swap for a CUDA build if available.
torch==2.3.1
torchvision==0.18.1

# Project A model runtime
vllm==0.5.4.post1

# Tooling / tests
pytest==8.2.0
pytest-asyncio==0.23.5
